requirements.txt
flask
flask-cors
pillow
imageio
imageio-ffmpeg
numpy
